requirements.txt
dbt-bigquery
google-cloud-bigquery
google-auth
pandas
pyarrow